// src/pages/enums.ts
export enum PaniniNames {
  bread = "base.bread",
  cheese = "base.cheese",
  meat = "base.meat",
  dressing = "base.dressing",
  vegetables = "base.vegetables",
  egg = "extras.egg",
  spreads = "extras.spreads",
  serving = "extras.serving",
  topping = "extras.topping",
  sandwichName = "sandwichName",
  cutlery = "cutlery",
  napkins = "napkins",
}